+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Added: tools made from Thermal Foundation (TF) materials
 * Added: config option to disable all tools made from TF materials
 * Added: shift + click ingredients into the table
+* Tested the build on a server BEFORE releasing -.-
 
 1.1.4
 * Fixed: Tables now work properly with Tinker's tools (#3)